Made it so Reservation only has a guestID, instead of a Class Guest. Also TableID. IDs now are longs instead of int.
Started with GetCompleteDB, which now returns lists of tables, guest, and reservations.

+++ Data/Guest.cs
@@ -8,12 +8,12 @@ namespace Data
 {
     public class Guest
     {
-        public int ID;
+        public long ID;
         public string Name;
         public string TelephoneNumber;
         public string EmailAddress;
 
-        public Guest(int id, string name, string telephoneNumber, string emailAddress)
+        public Guest(long id, string name, string telephoneNumber, string emailAddress)
         {
             ID = id;
             Name = name;
